refactor: remove BundledUpdate/checkNotInMainThread typealiases, use commons directly
These utilities were already moved to commons but amethyst still imported
them through typealias wrappers. This commit:

- Updates all BundledUpdate/BundledInsert imports to use commons.service directly
- Updates all checkNotInMainThread imports to use commons.threading directly
- Removes the now-unnecessary BundledUpdates.kt typealias file
- Removes the now-unnecessary MainThreadChecker.kt wrapper

This cleanup reduces indirection and unblocks ViewModel migration by
ensuring these commonly-used utilities are clearly in the commons module.
