+++ amethyst/src/play/java/com/vitorpamplona/amethyst/service/lang/LanguageTranslatorService.kt
@@ -30,7 +30,7 @@ import com.google.mlkit.nl.translate.TranslateLanguage
 import com.google.mlkit.nl.translate.Translation
 import com.google.mlkit.nl.translate.Translator
 import com.google.mlkit.nl.translate.TranslatorOptions
-import com.vitorpamplona.amethyst.service.checkNotInMainThread
+import com.vitorpamplona.amethyst.commons.threading.checkNotInMainThread
 import com.vitorpamplona.quartz.utils.urldetector.detection.UrlDetector
 import kotlinx.coroutines.CancellationException
 import java.util.concurrent.ExecutorService
